Refactor retrieval into two - retrieve web and retrieve doc
