+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "tsx index.ts",
-    "retrieve-doc": "tsx src/retrieval-chain.ts",
+    "retrieve-web": "tsx src/retrieval-chain/web-answer.ts",
+    "retrieve-doc": "tsx src/retrieval-chain/doc-answer.ts",
     "dev": "nodemon -w ./index.ts --exec tsx index.ts"
   },
   "author": "",
